Move class out of main.
